Add support for web analytics using Piwik, fixes #855
Does not do anything unless `public.piwik` setting is populated with the
following properties:

```
"piwik": {
    "url": "http://piwik.example.com/",
    "site": 1
}
```

Where `url` is the base url to the piwik instance (terminating slash is
important) and `site` is the numeric id of the site configured on that
piwik instance.

Iron-router integration was implemented in accordance with
recommendations for single page applications published in the
[piwik blog][1].

Regrettably there is currently a [piwik bug][2] preventing reliable
user-tracking in single-page-applications, hence this is not supported
at this time.

[1]: https://piwik.org/blog/2017/02/how-to-track-single-page-websites-using-piwik-analytics/
[2]: https://github.com/piwik/piwik/issues/7556